feat: Add assignees for each issue to the sprint report
Each assignee is appended to the end of each issue if `--skip-names` is unset.
The assignee name is automaticaly condensed down into the smallest unique representation of:
 - first name only
 - first name, surname initial
 - full name
Where we assume the first name is the first word in the full name, and the surname is all other words.

Includes minor refactoring where each issue is a JiraIssue object, rather than dictionary, to simplify some field fetching.
Also includes removal of __pycache__ in .gitignore